+++ doc/whatsnew.rst
@@ -48,6 +48,10 @@ v0.2.8.dev
 
 - Add power mean estimation in :func:`pyriemann.utils.mean.mean_power`
 
+- Add example to compare classifiers on synthetic datasets
+
+- Add ``predict_proba`` in :class:`pyriemann.classification.KNearestNeighbor`, and correct attribute ``classes_``
+
 
 v0.2.7 (June 2021)
 ------------------